requirements.txt
aiohttp~=3.7.4.post0
async-timeout~=3.0.1
discord.py~=2.0.0
dnspython~=2.1.0
jishaku==2.4.0
motor==3.0.0
pymongo==4.1.1
python-dotenv~=0.20.0
pytz~=2021.3
googletrans~=4.0.0rc1
asyncpg~=0.25.0
psutil~=5.9.1
